requirements.txt
torch>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
duckdb>=0.10.2
SQLAlchemy>=2.0.0
pandas>=2.0.0
pyarrow>=12.0.0
flask>=2.3.0
flask-cors>=4.0.0
plotly>=5.14.0
matplotlib>=3.7.0
seaborn>=0.12.0
pyyaml>=6.0
python-dotenv>=1.0.0
tqdm>=4.65.0
colorama>=0.4.6
python-dateutil>=2.8.2
pytest>=7.3.0
pytest-cov>=4.1.0
pytest-mock>=3.10.0
loguru>=0.7.0
joblib>=1.2.0
numba>=0.57.0
psutil>=5.9.0

# Optional: Experiment Tracking
# mlflow>=2.3.0

# Development (optional)
# black>=23.3.0
# isort>=5.12.0
# flake8>=6.0.0
# pylint>=2.17.0
# mypy>=1.3.0
